+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ### Added
 - Checkbox component
 
+### Changed
+- Add `min-width` and `text-align` for Dropdown's button.
+- Add `isFluid` prop to `Dropdown` and `ToggleMenu` components.
+
 ## [0.8.2] - 2018-09-27
 ### Changed
 - Fix classnames in Modal component